project_fetal/requirements.txt
streamlit 
numpy
tensorflow
keras.models
PIL